fix: abstract out components
