+++ app/components/About/Bio.tsx
