+++ app/page.tsx
@@ -1,4 +1,4 @@
-import About from "./about";
+import About from "./components/About";
 
 export default async function Home() {
   return (
